pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "firecrawl-dashboard"
version = "0.1.0"
description = "Web monitoring dashboard for self-hosted Firecrawl instances"
readme = "docs/README.md"
requires-python = ">=3.12"
authors = [
    {name = "Christian", email = "christian@example.com"},
]
keywords = ["firecrawl", "monitoring", "dashboard", "web-scraping"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aiohttp>=3.11.18",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "redis>=6.1.0",
    "uvicorn>=0.34.2",
]

[project.urls]
Homepage = "https://github.com/your-username/firecrawl-dashboard"
Documentation = "https://github.com/your-username/firecrawl-dashboard/blob/main/docs/README.md"
Repository = "https://github.com/your-username/firecrawl-dashboard.git"
Issues = "https://github.com/your-username/firecrawl-dashboard/issues"

[project.scripts]
firecrawl-dashboard = "firecrawl_dashboard.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/firecrawl_dashboard"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
